test: migrate part of DatePicker tests (#37115)
* test: migrate part of DatePicker tests

* test: migrate part of DatePicker tests

* fix: update snap

* test: migrate part of DatePicker tests

* test: migrate part of DatePicker tests

* test: migrate part of DatePicker tests

* test: migrate part of DatePicker tests

* test: migrate part of DatePicker tests

* test: migrate part of DatePicker tests

* test: remove useless methods

* test: add test ignore
