+++ components/date-picker/__tests__/QuarterPicker.test.tsx
@@ -1,6 +1,6 @@
-import { mount } from 'enzyme';
 import React from 'react';
 import DatePicker from '..';
+import { render } from '../../../tests/utils';
 import { resetWarned } from '../../_util/warning';
 
 const { QuarterPicker } = DatePicker;
@@ -10,8 +10,8 @@ describe('QuarterPicker', () => {
     resetWarned();
     const warnSpy = jest.spyOn(console, 'error');
 
-    const wrapper = mount(<QuarterPicker style={{ width: 400 }} />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { container } = render(<QuarterPicker style={{ width: 400 }} />);
+    expect(container.firstChild).toMatchSnapshot();
 
     expect(warnSpy).toHaveBeenCalledWith(
       "Warning: [antd: QuarterPicker] DatePicker.QuarterPicker is legacy usage. Please use DatePicker[picker='quarter'] directly.",
