+++ components/date-picker/__tests__/WeekPicker.test.tsx
@@ -1,8 +1,7 @@
-import { mount } from 'enzyme';
 import React from 'react';
 import DatePicker from '..';
 import focusTest from '../../../tests/shared/focusTest';
-import { resetMockDate, setMockDate } from '../../../tests/utils';
+import { render, resetMockDate, setMockDate } from '../../../tests/utils';
 
 const { WeekPicker } = DatePicker;
 
@@ -18,7 +17,7 @@ describe('WeekPicker', () => {
   focusTest(WeekPicker, { refFocus: true });
 
   it('should support style prop', () => {
-    const wrapper = mount(<WeekPicker style={{ width: 400 }} />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { container } = render(<WeekPicker style={{ width: 400 }} />);
+    expect(container.firstChild).toMatchSnapshot();
   });
 });
