+++ components/date-picker/__tests__/other.test.tsx
@@ -1,12 +1,13 @@
-import { mount, render } from 'enzyme';
 import moment from 'moment';
 import React from 'react';
 import DatePicker from '..';
 import ConfigProvider from '../../config-provider';
+import type { Locale } from '../../locale-provider';
 import LocaleProvider from '../../locale-provider';
 import locale from '../../locale-provider/zh_CN';
 import jaJP from '../../locale/ja_JP';
 import zhTW from '../locale/zh_TW';
+import { render } from '../../../tests/utils';
 
 const { MonthPicker, WeekPicker } = DatePicker;
 
@@ -23,15 +24,14 @@ describe('Picker format by locale', () => {
   };
 
   const date = moment('2000-01-01', 'YYYY-MM-DD');
-  function matchPicker(name, Picker, props) {
+  function matchPicker(name: string, Picker: typeof MonthPicker | typeof WeekPicker, props?: any) {
     it(name, () => {
-      const wrapper = mount(
-        <LocaleProvider locale={myLocale}>
+      const { container } = render(
+        <LocaleProvider locale={myLocale as Locale}>
           <Picker value={date} {...props} />
         </LocaleProvider>,
       );
-
-      expect(wrapper.render()).toMatchSnapshot();
+      expect(container.firstChild).toMatchSnapshot();
     });
   }
 
@@ -44,35 +44,33 @@ describe('Picker format by locale', () => {
 describe('MonthPicker and WeekPicker', () => {
   it('render MonthPicker', () => {
     const birthday = moment('2000-01-01', 'YYYY-MM-DD').locale('zh-cn');
-    const wrapper = mount(<MonthPicker open />);
-    wrapper.setProps({ value: birthday });
-    expect(render(wrapper.find('Trigger').instance().getComponent())).toMatchSnapshot();
+    const { container } = render(<MonthPicker open value={birthday} />);
+    expect(container.querySelector('div.ant-picker-dropdown')?.parentNode).toMatchSnapshot();
   });
 
   it('render WeekPicker', () => {
     const birthday = moment('2000-01-01', 'YYYY-MM-DD').locale('zh-cn');
-    const wrapper = mount(<WeekPicker open />);
-    wrapper.setProps({ value: birthday });
-    expect(render(wrapper.find('Trigger').instance().getComponent())).toMatchSnapshot();
+    const { container } = render(<WeekPicker open value={birthday} />);
+    expect(container.querySelector('div.ant-picker-dropdown')?.parentNode).toMatchSnapshot();
   });
 });
 
 describe('Override locale setting of the ConfigProvider', () => {
   it('DatePicker', () => {
-    const wrapper = mount(
+    const { container } = render(
       <ConfigProvider locale={jaJP}>
         <DatePicker locale={zhTW} />
       </ConfigProvider>,
     );
-    expect(wrapper.find('input').props().placeholder).toEqual('請選擇日期');
+    expect(container.querySelector('input')?.placeholder).toEqual('請選擇日期');
   });
   it('RangePicker', () => {
-    const wrapper = mount(
+    const { container } = render(
       <ConfigProvider locale={jaJP}>
         <DatePicker.RangePicker locale={zhTW} />
       </ConfigProvider>,
     );
-    expect(wrapper.find('input').at(0).props().placeholder).toEqual('開始日期');
-    expect(wrapper.find('input').at(1).props().placeholder).toEqual('結束日期');
+    expect(container.querySelectorAll('input')[0]?.placeholder).toEqual('開始日期');
+    expect(container.querySelectorAll('input')[1]?.placeholder).toEqual('結束日期');
   });
 });
