+++ components/table/__tests__/Table.expand.test.js
@@ -81,7 +81,7 @@ describe('Table.expand', () => {
     fireEvent.click(container.querySelector('.ant-table-row-expand-icon'));
     expect(container.querySelector('.indent-level-1').style.paddingLeft).toEqual('0px');
   });
-  
+
   it('has right aria-expanded state', () => {
     const { container } = render(<Table columns={columns} dataSource={data} />);
     expect(container.querySelector('[aria-expanded=false]')).toBeTruthy();
